refactor: codebase -> module dreambooth_keras + added packaging info
