+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "restorers"
+name = "dreambooth_keras"
 dynamic = ["version"]
 description = "Implementation of DreamBooth in KerasCV and TensorFlow."
 readme = "README.md"
@@ -15,12 +15,16 @@ classifiers = [
 
 dependencies = [
     "keras_cv==0.4.0",
+    "tensorflow>=2.10.0",
     "tensorflow_datasets>=4.8.1",
     "pillow==9.4.0",
     "wandb>=0.13.9",
     "imutils",
     "opencv-python",
 ]
+
+[tool.setuptools]
+packages = ["dreambooth_keras"]
 
 [project.urls]
 homepage = "https://github.com/sayakpaul/dreambooth-keras"
